Color scheme changes and minor bug fixes
